requirements.txt
sqlalchemy
requests
python-dotenv
psycopg2-binary